Refactoring, adjustments
